ENH: Add a new parameter fclear_lim to compute snow line with enough clear pixels
fclear_lim is the minimum percentage of clear pixels in an elevation band
(default value 0.1) used to compute the snow line.

It avoids to compute snow line with limited numbers of pixels in an elevation
brands wich will can bias the 'zs' computation and then the snow mask computed
at the pass 2 with a less strict threshold on the NDSI.

Clean also code related to deprecated functions and related test in the C++ part
(compute_zs and compute_snow_line)

The ATBD has been modified to document the

By default older configuration version will continue to work with a default
value of 0.1 for fclear_lim.